added middleware and fixed some bug and api for login recruiter
1-> Added recruiter auth middleware .
2-> Design an api for recuiter login
3-> fixed some bug regarding cloudinary and restrict the file selection to image and pdf for profile image and resume respectively.
4-> fixed some bug

+++ server/src/utils/cloudinary.js
@@ -1,9 +1,9 @@
-
-// import { v2 as cloudinary } from 'cloudinary' 
-const fs = require("fs")
-const cloud = require("cloudinary")
+// import { v2 as cloudinary } from 'cloudinary'
+const fs = require("fs");
+const cloud = require("cloudinary");
 const cloudinary = cloud.v2;
-require('dotenv').config();
+require("dotenv").config();
+const path = require("path");
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -17,39 +17,47 @@ cloudinary.config({
 
 const uploadOnCloudinary = async (localFilePath) => {
   try {
-      if (!localFilePath) return null
-      //upload the file on cloudinary
-      const response = await cloudinary.uploader.upload(localFilePath, {
-          resource_type: "auto"
-      })
-      // file has been uploaded successfull
-      console.log("file is uploaded on cloudinary ", response.url);
-      fs.unlinkSync(localFilePath)
-      return response;
+    if (!localFilePath) return null;
+    //upload the file on cloudinary
+    const ext = path.extname(localFilePath).toLowerCase();
+    const allowedExtensions = [".jpg", ".jpeg", ".png", ".webp"];
+    if (!allowedExtensions.includes(ext)) {
+      throw new Error("Only image files are allowed");
+    }
 
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "image",
+    });
+    // file has been uploaded successfull
+    console.log("file is uploaded on cloudinary ", response.url);
+    fs.unlinkSync(localFilePath);
+    return response;
   } catch (error) {
-      fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-      return null;
+    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+    throw error;
   }
-}
+};
 
 const uploadPdfOnCloudinary = async (localFilePath) => {
   try {
-      if (!localFilePath) return null
-      //upload the file on cloudinary
-      const response = await cloudinary.uploader.upload(localFilePath, {
-          resource_type: "raw"
-      })
-      // file has been uploaded successfull
-      console.log("file is uploaded on cloudinary ", response.url);
-      fs.unlinkSync(localFilePath)
-      return response;
+    if (!localFilePath) return null;
+    //upload the file on cloudinary
 
+    const ext = path.extname(localFilePath).toLowerCase();
+    if (ext !== ".pdf") {
+      throw new Error("Only Pdf file allowed");
+    }
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
+    // file has been uploaded successfull
+    console.log("file is uploaded on cloudinary ", response.url);
+    fs.unlinkSync(localFilePath);
+    return response;
   } catch (error) {
-      fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-      return null;
+    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+    throw error;
   }
-}
+};
 
-
- module.exports ={uploadOnCloudinary,uploadPdfOnCloudinary}
+module.exports = { uploadOnCloudinary, uploadPdfOnCloudinary };
